Intel(R) SGX DCAP 1.20 Release
Introduced the Intel DCAP Appraisal Engine within quote verification library,
  empowering users to evaluate verification results against diverse policies.
Upgraded Intel SGX Quote Verification Enclave to integrate OpenSSL/SgxSSL 3.0.12.
Added Rust wrapper for quote provider library APIs.
Fixed bugs.

+++ tools/SGXPlatformRegistration/common/src/common.cpp
@@ -124,7 +124,7 @@ int readFileToBuffer(const char *filename, uint8_t *buffer, size_t &buffSize) {
         goto out;
     }
 
-    writtenSize = fread(buffer, sizeof(uint8_t), fileSize, file);
+    writtenSize = fread(buffer, sizeof(uint8_t), (unsigned long)fileSize, file);
     if (writtenSize != (unsigned long)fileSize) {
         //agent_log_message(MP_REG_LOG_LEVEL_ERROR, "Failed to read data from file: %s\n", filename);
         ret = MP_UNEXPECTED_ERROR;
